\name{TwinEQTL-package}
\alias{TwinEQTL}
\docType{package}
\title{
        TwinEQTL: Fast and Powerful Association Analysis 
        for eQTL and GWAS in Twin Studies
}
\description{
        A computationally efficient alternative,
        to linear mixed-effects model (LMM) for twin
        genome-wide association study (GWAS) and 
        expression quantitative trait loci (eQTL) analyses.
}
\author{
    Andrey A Shabalin \email{andrey.shabalin@gmail.com}

	Maintainer: Andrey A Shabalin <andrey.shabalin@gmail.com>
}
\references{
	The package website: \url{https://github.com/andreyshabalin/TwinEQTL}
}
\keyword{ package }
\keyword{ TwinEQTL }
\seealso{
	See \code{\link{TwinEQTL_testAll}} for reference and sample code.
}
