\name{TwinMeta-package}
\alias{TwinMeta}
\docType{package}
\title{
        TwinMeta: Fast and Powerful Association Analysis 
        for eQTL and GWAS in Twin Studies
}
\description{
        A computationally efficient alternative,
        to linear mixed-effects model (LMM) for twin
        genome-wide association study (GWAS) and 
        expression quantitative trait loci (eQTL) analyses.
}
\details{
	\tabular{ll}{
		Package: \tab TwinMeta\cr
		Type: \tab Package\cr
		Version: \tab 1.0\cr
		Date: \tab 2021-07-20\cr
		License: \tab LGPL-3 \cr
		LazyLoad: \tab yes\cr
		Depends: \tab MatrixEQTL, methods, utils, stats\cr
	}
}
\author{
        Andrey A Shabalin \email{andrey.shabalin@gmail.com}

	Maintainer: Andrey A Shabalin <andrey.shabalin@gmail.com>
}
\references{
	The package website: \url{https://github.com/andreyshabalin/TwinMeta}
}
\keyword{ package }
\keyword{ TwinMeta }
\seealso{
	See \code{\link{TwinMeta_testAll}} for reference and sample code.
}
